Update the arrows to be closer to Figma
## What
I updated the project to include font awesome which has a lot of arrows that would work. https://fontawesome.com/search?q=arrow&o=r&m=free&c=arrows

Unfortunately, the one in the figma involves arrows which cost money (pro) and layering so I couldn't go that far, but his one is a single glyph which mostly gets the job done with less pain in the css. Going to just recommend we keep this unless there's a reason to work on the css layering for better testimonial arrows.